Added all design. Fixed bugs.

+++ Grade3(2022-2023)/OOP/Semester2/lab1_server/src/main/java/com/example/servlets/SetClientPhoneServlet.java
@@ -31,7 +31,7 @@ public class SetClientPhoneServlet extends DBServlet {
 
         try {
             int clientId = requestJsonData.getInt("clientId");
-            int clientPhone = requestJsonData.getInt("clientPhone");
+            long clientPhone = requestJsonData.getLong("clientPhone");
             dbManager.SetClientPhone(clientId, clientPhone);
             out.println("true");
             logger.info("SET CLIENT WITH ID " + clientId +" PHONE NUMBER "+clientPhone);
